Do not kill background processes
If the cockpit-bridge session is terminated every child process receives
the signal and is killed. If a exec() is called with a "stream" argument
(so that background process tracking is requested), we spawn the process
in a separate process group, so it can detach from the cockpit session
and survive to session cut off.